Move helper files to helper directories
