+++ frontend/test/metabase/scenarios/binning/correctness/longitude.cy.spec.js
@@ -5,7 +5,7 @@ import {
   summarize,
 } from "__support__/e2e/helpers";
 
-import { LONGITUDE_OPTIONS } from "./constants";
+import { LONGITUDE_OPTIONS } from "./helpers/constants";
 
 describe("scenarios > binning > correctness > longitude", () => {
   beforeEach(() => {
