+++ frontend/test/metabase/scenarios/cross-version/source/helpers/cross-version-source.config.js
@@ -1,6 +1,6 @@
 const { defineConfig } = require("cypress");
 const {
   crossVersionSourceConfig,
-} = require("../../../../__support__/e2e/config");
+} = require("../../../../../__support__/e2e/config");
 
 module.exports = defineConfig({ e2e: crossVersionSourceConfig });
