+++ frontend/test/metabase/scenarios/dashboard-filters/helpers/dashboard-filters-sql-text-category.js
