+++ frontend/test/metabase/scenarios/dashboard-filters/helpers/dashboard-filters-text-category.js
