+++ frontend/test/metabase/scenarios/embedding/helpers/embedding-dashboard.js
