+++ frontend/test/metabase/scenarios/embedding/helpers/embedding-snippets.js
